Clean up docs on C++ interface
Rephrase the introduction to clarify the differences between the C++ program and the Soufflé program.

Number sections from 1 instead of 0.

Add a main function and `#include` to the complete example. Since most users of the C++ interface will probably be using their own main function this should be more helpful.